+++ examples/example/spec/second_spec.rb
@@ -1,12 +1,12 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'second' do
-
-  it "works" do
+  it 'works' do
     browser.navigate.to page 1
-    element = find(data: "good div")
+    element = find(data: 'good div')
     browser.execute_script("arguments[0].setAttribute('id', 'lost')", element)
-    expect(find(data: "good div").text).to eq "Text"
+    expect(find(data: 'good div').text).to eq 'Text'
   end
-
 end